split PalletBuilderViewModel into 3 viewmodels for each tab
