+++ App.xaml.cs
@@ -4,6 +4,7 @@ using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Stack_Solver.Data;
 using Stack_Solver.Data.Repositories;
+using Stack_Solver.Infrastructure;
 using Stack_Solver.Services;
 using Stack_Solver.ViewModels.Pages;
 using Stack_Solver.ViewModels.Windows;
@@ -60,6 +61,8 @@ namespace Stack_Solver
                 services.AddSingleton<TruckLoadingViewModel>();
                 services.AddSingleton<JobManagerPage>();
                 services.AddSingleton<JobManagerViewModel>();
+
+                services.AddSingleton<IEventAggregator, EventAggregator>();
 
                 services.AddDbContextFactory<ApplicationDbContext>(options =>
                 {
